sales-call-transcription-ai/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Audio processing
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0

# Speaker diarization
pyannote.audio==3.1.1

# YouTube download
yt-dlp==2024.1.28

# Audio format handling
pydub==0.25.1
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
